# docs/assets/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Token do painel": {
        "description": "Perfil local: o valor de `SDR_PAINEL_TOKEN` (em desenvolvimento, `dev-token`). Perfil AWS: o JWT emitido pelo Cognito.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Projeto Mora — Vértice Imóveis (POC acadêmica)"
    },
    "description": "\nAPI do **Mora**, o agente SDR imobiliário da Vértice Imóveis.\n\nA etiqueta de cada rota diz quem a usa e o que ela exige:\n\n| Etiqueta | Quem usa | Autenticação |\n|---|---|---|\n| `público` | site vitrine e chat | nenhuma |\n| `corretor` | painel do corretor | `Authorization: Bearer …` |\n| `operacao` | rotinas de atendimento no painel | `Authorization: Bearer …` |\n| `admin` | configuração, governança de IA e auditoria | `Authorization: Bearer …` |\n\n**Como autenticar no perfil local:** clique em **Authorize** e informe o valor de\n`SDR_PAINEL_TOKEN` (no ambiente de desenvolvimento, `dev-token`). No perfil AWS o token é o\nJWT do Cognito, validado pelo authorizer do API Gateway antes de chegar aqui.\n\nToda alteração feita por estas rotas é registrada em auditoria (ver `GET /auditoria`).\n",
    "license": {
      "name": "Uso acadêmico — licença a confirmar"
    },
    "title": "Mora — SDR Imobiliário API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auditoria": {
      "get": {
        "operationId": "listar_auditoria_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "ator",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ator"
            }
          },
          {
            "in": "query",
            "name": "acao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acao"
            }
          },
          {
            "in": "query",
            "name": "entidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entidade"
            }
          },
          {
            "in": "query",
            "name": "entidade_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entidade Id"
            }
          },
          {
            "in": "query",
            "name": "busca",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          },
          {
            "in": "query",
            "name": "so_sensiveis",
            "required": false,
            "schema": {
              "default": false,
              "title": "So Sensiveis",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 300,
              "maximum": 1000,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "admin"
        ]
      }
    },
    "/auditoria/exportar": {
      "get": {
        "description": "CSV do período — para enviar a quem pediu a prestação de contas.",
        "operationId": "exportar_auditoria_exportar_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "acao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acao"
            }
          },
          {
            "in": "query",
            "name": "entidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entidade"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Exportar",
        "tags": [
          "admin"
        ]
      }
    },
    "/calendario/callback": {
      "get": {
        "description": "Volta do Google. Endpoint público por definição — quem chama é o navegador, sem o nosso token.\nO `state` é assinado por nós (ver seguranca/oauth.py): sem isso, bastava adivinhar o id de um\ncorretor para plantar a credencial da PRÓPRIA conta Google na agenda dele.",
        "operationId": "callback_calendario_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "admin"
        ]
      }
    },
    "/calendario/conectar/{corretor_id}": {
      "post": {
        "description": "Devolve a URL de consentimento. Quem abre é o navegador do corretor, não a API.",
        "operationId": "conectar_calendario_conectar__corretor_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "corretor_id",
            "required": true,
            "schema": {
              "title": "Corretor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Conectar",
        "tags": [
          "admin"
        ]
      }
    },
    "/calendario/status": {
      "get": {
        "description": "O painel usa isto para saber quem já conectou. Nenhum token sai daqui.",
        "operationId": "status_calendario_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/calendario/{corretor_id}": {
      "delete": {
        "operationId": "desconectar_calendario__corretor_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "corretor_id",
            "required": true,
            "schema": {
              "title": "Corretor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Desconectar",
        "tags": [
          "admin"
        ]
      }
    },
    "/clientes": {
      "get": {
        "operationId": "listar_clientes_get",
        "parameters": [
          {
            "in": "query",
            "name": "busca",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Busca"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "title": "Limite",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Listar",
        "tags": [
          "operacao"
        ]
      }
    },
    "/clientes/{cliente_id}": {
      "get": {
        "operationId": "ficha_clientes__cliente_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_id",
            "required": true,
            "schema": {
              "title": "Cliente Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Token do painel": []
          }
        ],
        "summary": "Ficha",
        "tags": [
          "operacao"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Perfil local (docker compose)",
      "url": "http://localhost:8000"
    },
    {
      "description": "Perfil AWS — a confirmar no deploy",
      "url": "https://api.exemplo.com.br"
    }
  ],
...
}